feat: 添加餐厅详情居中卡片视图和交互优化
- 新增 CenteredDetailCardView 居中详情卡片组件
- 新增 ExpandedRestaurantView 展开内容组件
- 餐厅评分从 Int 改为 Double 支持半星评分
- 详情页使用 iOS 原生 Menu 选择品类
- 详情页使用内联可点击星标评分
- 详情页封面图支持更换
- 详情页支持二次编辑品类、评分、印象
- 底部导航条在详情页自动隐藏
- 打卡按钮移到标题右侧
- 详情卡片高度增加到屏幕宽度的2倍
- 列表卡片星级显示整数
- ContentView 添加 tabBar 隐藏/恢复通知

+++ WhatToEat/Restaurant.swift
@@ -8,20 +8,19 @@ final class Restaurant {
     var type: String
     var district: String
     var city: String
-    var rating: Int
+    var rating: Double
     var address: String
     var latitude: Double
     var longitude: Double
-    var coverPhotoFilename: String?  // 只能存字符串文件名
+    var coverPhotoFilename: String?
     var review: String
     var tags: [String]
     var averagePrice: Double
     var createdAt: Date
     
-    // 关系：确保这里写得标准
     @Relationship(deleteRule: .cascade) var logs: [VisitLog] = []
     
-    init(name: String = "", type: String = "未分类", district: String = "其他", city: String = "", rating: Int = 3, address: String = "", latitude: Double = 0.0, longitude: Double = 0.0, coverPhotoFilename: String? = nil, review: String = "", tags: [String] = [], averagePrice: Double = 0.0) {
+    init(name: String = "", type: String = "未分类", district: String = "其他", city: String = "", rating: Double = 3.0, address: String = "", latitude: Double = 0.0, longitude: Double = 0.0, coverPhotoFilename: String? = nil, review: String = "", tags: [String] = [], averagePrice: Double = 0.0) {
         self.id = UUID()
         self.name = name
         self.type = type
